Refactoring using github.com/corelayer/clapp
