+++ go.mod
@@ -1,10 +1,28 @@
 module github.com/corelayer/netscaleradc-acme-go
 
-go 1.20
+go 1.21
 
 require (
-	github.com/corelayer/netscaleradc-nitro-go v0.0.0-20230509103033-105855723b63 // indirect
+	github.com/corelayer/clapp v0.0.0-20230825151738-4c9ff029be46
+	github.com/corelayer/netscaleradc-nitro-go v0.0.0-20230509103033-105855723b63
+	github.com/spf13/cobra v1.7.0
+	github.com/spf13/viper v1.16.0
+)
+
+require (
+	github.com/fsnotify/fsnotify v1.6.0 // indirect
+	github.com/hashicorp/hcl v1.0.0 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/cobra v1.7.0 // indirect
+	github.com/magiconair/properties v1.8.7 // indirect
+	github.com/mitchellh/mapstructure v1.5.0 // indirect
+	github.com/pelletier/go-toml/v2 v2.0.9 // indirect
+	github.com/spf13/afero v1.9.5 // indirect
+	github.com/spf13/cast v1.5.1 // indirect
+	github.com/spf13/jwalterweatherman v1.1.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
+	github.com/subosito/gotenv v1.6.0 // indirect
+	golang.org/x/sys v0.11.0 // indirect
+	golang.org/x/text v0.12.0 // indirect
+	gopkg.in/ini.v1 v1.67.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
